+++ fluxpype/config.ini
@@ -1,14 +1,10 @@
 [DEFAULT]
 config_name = velocityplots  # Name of the configuration to be used by default
-
 jsoc_email = chris.gilly@colorado.edu  # Email to use for JSOC data queries
+base_dir = ~/vscode/fluxons # Directory in which fluxon-MHD resides
+run_script = fluxpype/fluxpype/science/magnetogram2wind.pdl  # Path to the run script for the pipeline
 
 # Base Paths
-; run_script = ~/fluxpype/fluxpype/science/magnetogram2wind.pdl
-base_dir = ~/vscode/fluxons
-run_script = fluxpype/science/magnetogram2wind.pdl  # Path to the run script for the pipeline
-; run_script = 23
-# Paths (use base_dir for dynamic reference)
 rc_path = ~/.zshrc  # Location of .zshrc file for shell configuration
 python_dir = /fluxpype/.venv/bin/python3  # Path to Python executable
 pl_prefix = ~/Library/perl5  # Perl scripts location prefix
@@ -18,7 +14,6 @@ data_dir = ${base_dir}/fluxpype/fluxpype/data  # Location where fluxon-data is s
 working_dir = ${base_dir}/fluxpype
 nwant = 1
 n_jobs = 1
-; cr = 2200
 
 #############################################
 # Default Batch parameters
@@ -73,11 +68,14 @@ adapts = [0, 1]
 adapt_select = 2
 
 [velocityplots]
-batch_name = Newtest
-rotations  = [2225]  #, 2250, 5) #(2190, 2230, 1) # (start, stop, step) when parens are used
-fluxon_count = [1000]
+batch_name = Newtest5
+rotations  = [2260]  #, 2250, 5) #(2190, 2230, 1) # (start, stop, step) when parens are used
+fluxon_count = [1001]
+mag_reduce = 4 # Factor by which to reduce magnetic field data
+
 # Wind Modeling parameters
 flow_method = [parker] #, wsa, ghosts, parker, cranmer, tempest] #, ghosts, wsa, parker] # can be [tempest, ghosts, wsa, parker, cranmer]
-concurrency = 14  # Number of concurrent processes to run
+concurrency = 20  # Number of concurrent processes to run
 recompute_wind = 0 # Force re-computation of existing wind data (0 for off, 1 for on)
 wind_only = 0   # Skip fluxon calculation (0 for off, 1 for on)
+do_wind_calc = 1  # Calculate solar wind (0 for off, 1 for on)
