+++ fluxpype/pipe_helper.py
@@ -55,7 +55,7 @@ import pipe_helper as ph
 result = ph.convert_value("42")
 
 # Example usage of the configurations module
-from pipe_helper import configurations
+from fluxpype.pipe_helper import configurations
 configs = configurations()
 
 
@@ -80,7 +80,7 @@ from pathlib import PosixPath, Path
 import pandas as pd
 import re
 
-# from pipe_helper import convert_value
+# from fluxpype.pipe_helper import convert_value
 import numpy as np
 import matplotlib.pyplot as plt
 
@@ -619,6 +619,7 @@ def reduce_mag_file(mag_file, reduction=3, force=False):
     small_file = PosixPath(str(mag_file).replace("_r1_", f"_r{reduction}_"))
     # reduce the FITS image
     print(f"\tReducing image size by a factor of {reduction}...", end="")
+
     if not os.path.exists(small_file) or force:
         small_file = reduce_fits_image(
             mag_file, small_file, target_resolution=None, reduction_amount=reduction
@@ -666,7 +667,7 @@ def reduce_fits_image(
             data = hdul[1].data
 
         current_resolution = max(data.shape)
-        print("\t\tOriginal Shape: ", data.shape)
+        print("\t\tOriginal Shape:  ", data.shape)
 
         # Calculate the reduction amount if target resolution is specified
         if target_resolution is not None:
@@ -705,7 +706,7 @@ def reduce_fits_image(
             360 / (small_image.shape[0] * np.pi)
         )  # RADIANS
 
-        print("\t\tFinal Shape:    ", small_image.shape)
+        print("\t\tFinal Shape:     ", small_image.shape)
 
         print("\tSaving  ", small_file)
         fits.writeto(small_file, small_image, useheader, overwrite=True)
@@ -1054,7 +1055,7 @@ def add_paths(flux_pipe_dir):
 
 
 def make_mag_dir(datdir):
-    mag_dir = os.path.join(datdir, "magnetograms")
+    mag_dir = os.path.expanduser(os.path.join(datdir, "magnetograms"))
 
     if not os.path.exists(mag_dir):
         os.makedirs(mag_dir)
@@ -1116,6 +1117,8 @@ def get_magnetogram_file(
     found_file = False
     inst = "hmi"
     for file in file_list:
+        file = str(file.expanduser())
+        # print("\t\t\t", file)
         if str(CR) + "_r1_" in str(file) and inst in str(file).casefold():
             print(f"\t\tFound '{os.path.basename(file)}' in '{shorten_path(mag_dir)}'")
             found_file = True
@@ -1127,7 +1130,7 @@ def get_magnetogram_file(
         else:
             small_path = reduce_mag_file(file, reduce, force=force_download)
             return file, small_path
-
+    print("\t\tFile not found...")
     c = drms.Client()
     # Generate a search
     crot = a.jsoc.PrimeKey("CAR_ROT", str(CR))
@@ -1395,7 +1398,7 @@ def reduce_mag_file(mag_file, reduction=3, force=False):
         # print("\t\t", shorten_path(str(small_file), 2))
         ### WORKING HERE
         print(
-            f"\t\tFound '{os.path.basename(small_file)}' in '{shorten_path(os.path.dirname(small_file))}'"
+            f"\t\tFound '{os.path.basename(small_file)}' in '{os.path.dirname(small_file)}'"
         )
         print("\n\t\t\t```````````````````````````````\n \n\n")
 
@@ -1480,7 +1483,7 @@ def reduce_fits_image(
         # useheader['CDELT1'] = 360 / small_image.shape[1]  ## DEGREES
         # useheader['CDELT2'] = np.deg2rad(360 / (small_image.shape[0] * np.pi)) #RADIANS
 
-        print("\tFinal Shape:    ", small_image.shape)
+        print("\tFinal Shape: ", small_image.shape)
 
         print("\tSaving  ", small_file)
         fits.writeto(small_file, small_image, useheader, overwrite=True)
@@ -1819,3 +1822,149 @@ def parse_big_dict(big_dict, field="vel1", vmin=VMIN, vmax=VMAX):
         total_std,
         all_count,
     )
+
+
+
+# A quick function to read output FLUX world data
+
+# Import libraries
+import numpy
+
+def rdworld(filename):
+
+    # Some quick testing on storing this data in a class object
+    class world(object):
+
+        def __init__(self):
+            self.fc = self.fc()
+            self.fx = self.fx()
+
+        # Flux concentrations
+        class fc:
+            def __init__(self):
+                self.id = []
+                self.x = []
+                self.y = []
+                self.z = []
+                self.fl = []
+
+        def add_fc(self, id, x, y, z, fl):
+            self.fc.id.append(id)
+            self.fc.x.append(x)
+            self.fc.y.append(y)
+            self.fc.z.append(z)
+            self.fc.fl.append(fl)
+
+        # Fluxons
+        class fx:
+            def __init__(self):
+                self.id = []
+                self.fc0 = []
+                self.fc1 = []
+                self.fl = []
+                self.x = []
+                self.y = []
+                self.z = []
+
+        def add_fx(self, id, fc0, fc1, fl, r0, r1, x, y, z):
+            self.fx.id.append(id)
+            self.fx.fc0.append(fc0)
+            self.fx.fc1.append(fc1)
+            self.fx.fl.append(fl)
+            self.fx.x.append(numpy.array([r0[0]] + x + [r1[0]]))
+            self.fx.y.append(numpy.array([r0[1]] + y + [r1[1]]))
+            self.fx.z.append(numpy.array([r0[2]] + z + [r1[2]]))
+
+    # Let there be light
+    w = world()
+
+    # Now actually go about reading the data
+    df = open(filename, 'r')
+
+    # Initialize data storage
+    ln_id = []
+    ln_fc0 = []
+    ln_fc1 = []
+    ln_fl = []
+    ln_r0 = []
+    ln_r1 = []
+
+    vx_lid = []
+    vx_id = []
+    vx_pos = []
+    vx_x = []
+    vx_y = []
+    vx_z = []
+
+    for rl in df:
+        sl = rl.strip().split()
+
+        # Skip ahead for blank lines
+        if len(sl) == 0: continue
+
+        # Read out flux concentrations
+        if sl[0] == 'NEW':
+            w.add_fc(numpy.int(sl[1]), numpy.double(sl[2]), numpy.double(sl[3]), numpy.double(sl[4]), numpy.double(sl[5]))
+
+        # Read out fluxon data
+        if sl[0] == 'LINE':
+            if numpy.int(sl[1])>0:
+                ln_id.append(numpy.int(sl[1]))
+                ln_fc0.append(numpy.int(sl[4]))
+                ln_fc1.append(numpy.int(sl[5]))
+                ln_fl.append(numpy.double(sl[6]))
+                ln_r0.append([numpy.double(sl[7]), numpy.double(sl[8]), numpy.double(sl[9])])
+                ln_r1.append([numpy.double(sl[10]), numpy.double(sl[11]), numpy.double(sl[12])])
+
+        # Read out vertex points
+        if sl[0] == 'VERTEX':
+            vx_lid.append(numpy.int(sl[1]))
+            vx_id.append(numpy.int(sl[2]))
+            vx_pos.append(numpy.int(sl[3]))
+            vx_x.append(numpy.double(sl[4]))
+            vx_y.append(numpy.double(sl[5]))
+            vx_z.append(numpy.double(sl[6]))
+
+        # Exit on neighbor information
+        if 'VNEIGHBOR' in sl[0]:
+            break
+
+    df.close()
+
+    # Convert lists into numpy arrays
+    ln_id  = numpy.array(ln_id)
+    ln_fc0 = numpy.array(ln_fc0)
+    ln_fc1 = numpy.array(ln_fc1)
+    ln_fl = numpy.array(ln_fl)
+    ln_r0 = numpy.array(ln_r0)
+    ln_r1 = numpy.array(ln_r1)
+
+    vx_lid = numpy.array(vx_lid)
+    vx_id = numpy.array(vx_id)
+    vx_pos = numpy.array(vx_pos)
+    vx_x = numpy.array(vx_x)
+    vx_y = numpy.array(vx_y)
+    vx_z = numpy.array(vx_z)
+
+    # Parse the line and vertex lists and create fluxon objects
+    for lid in ln_id:
+        wl = numpy.where(ln_id == lid)[0]
+        wv = numpy.where(vx_lid == lid)[0]
+
+        w.add_fx(lid, ln_fc0[wl][0], ln_fc1[wl][0], ln_fl[wl][0], ln_r0[wl,:][0].tolist(), ln_r1[wl,:][0].tolist(), vx_x[wv].tolist(), vx_y[wv].tolist(), vx_z[wv].tolist())
+
+    return w
+
+# A quick comment on reading FLUX output world files:
+
+# Flux concentrations
+# NEW	100	0.314000	-0.059000	-0.948000	-1	(null)	0
+# FC    ID      X               Y               Z               Flux    BND PT  Rad
+
+# Line
+# LINE	101	-302	-301	-1	100	1            7.61951 -2.51477 -7.52456   0.314 -0.059 -0.948
+# Line  ID      Start   End     FC0     FC1     Flux        StX     StY     StZ     EndX    EndY    EndZ
+
+# Vertex
+# VERTEX	101	10650	1	7.203456	-2.377456	-7.113698
+# Vertex        LineID  VtxID   Pos     X               Y               Z
