+++ fluxpype/plotting/plot_bmag_fill.py
@@ -42,10 +42,10 @@ import matplotlib.pyplot as plt
 from matplotlib import gridspec
 import numpy as np
 import shutil
-from pipe_helper import load_fits_magnetogram, get_fixed_coords
+from fluxpype.pipe_helper import load_fits_magnetogram, get_fixed_coords
 import sunpy.coordinates
 import cv2
-from pipe_helper import configurations
+from fluxpype.pipe_helper import configurations
