+++ fluxpype/plotting/plot_wind_map_detailed.py
@@ -26,8 +26,15 @@ Functions:
 
 import numpy as np
 import matplotlib as mpl
-mpl.use("qt5agg")
-import matplotlib.pyplot as plt
+try:
+    mpl.use("qt5agg")
+    import matplotlib.pyplot as plt
+    fig = plt.figure()
+except ImportError as e:
+    print(e)
+    print("Loading agg instead")
+    mpl.use("agg")
+
 import argparse
 import os
