+++ fluxpype/science/make_world_from_pfss.pm
@@ -9,6 +9,15 @@ use strict;
 use warnings;
 use Exporter qw(import);
 our @EXPORT_OK = qw(make_world_from_pfss);
+
+
+use FindBin;
+use lib $FindBin::Bin;       # Adds the script's directory
+use lib "$FindBin::Bin/..";  # Adds the parent directory
+use lib "$FindBin::Bin/../plotting";  # Adds the neighbor's directory
+use lib "$FindBin::Bin/../..";  # Adds the grandparent directory
+
+
 use PDL::AutoLoader;
 use PDL;
 use PDL::Transform;
@@ -16,11 +25,18 @@ use PDL::NiceSlice;
 use PDL::Options;
 use Flux;
 use PDL::IO::Misc;
-# use make_world_sphere qw(make_world_sphere);
+use make_world_sphere qw(make_world_sphere);
 use File::Path;
 use Time::HiRes qw(clock_gettime);
 use File::Basename qw(fileparse);
 use File::Basename;
+
+use FindBin;
+use lib $FindBin::Bin;       # Adds the script's directory
+use lib "$FindBin::Bin/..";  # Adds the parent directory
+use lib "$FindBin::Bin/../plotting";  # Adds the neighbor's directory
+
+
 use pipe_helper qw(shorten_path);
 use plot_world qw(plot_world);
 use Flux::World    qw(read_world);
@@ -306,7 +322,6 @@ sub make_world_from_pfss {
         }
         print "\tDone! $open_count open fluxons generated.\n";
 
-
         ## Generate closed fieldlines  ###################################
         print "\n \n\tGenerating closed fluxons from \n\t\t$closed_name_short\n\t ";
         my ($cfln, $cfl) = rle($cflnum);
@@ -331,21 +346,15 @@ sub make_world_from_pfss {
         my $total_fluxons = $open_count + $closed_count;
         print "\n\t    Total fluxons created: $total_fluxons\n";
 
-
-
-
-
-
     ## Generate the world  ############################################
-        print "\n\tGenerating the World...\n \n";
+        print "\n\tGenerating the Initial World...\n \n";
         my $fbg = make_world_sphere(@flines, {rmax=>21.5});
         # my $fbg = make_world_sphere(@flines, {rmax=>215});
 
         $world = str2world($fbg);
 
-
         ## Save the initial world state  ###############################
-        print "\n\tSaving the World...";
+        print "\n\tSaving the Initial World...";
         my $flen = $oflnum->max() + $cflnum->max() + 4; #@flines+0;
         if (! -d $world_out_dir ) {mkpath($world_out_dir) or die "Failed to create directory: $world_out_dir $!\n";}
         $world->write_world($world_out_path);
@@ -354,16 +363,17 @@ sub make_world_from_pfss {
 
 
         ## Plot the initial World State ################################
+        print "\t Plotting the Initial World!\n";
         plot_world($world, $datdir, $batch_name, $CR, $reduction, $n_fluxons_wanted, $adapt, $force_make_world, $lim, $lim2, $configs, "initial");
 
 
     } else {
-        print "\n\n \tSkipped! World already exists:\n";
+        print "\n\n \tSkipped conversion! World already exists:\n";
         my $short_world_out_path = shorten_path($world_out_path);
         print "\t    $short_world_out_path\n";
 
-        print "\t Plotting World!\n";
         my $world = read_world($world_out_path);
+
         plot_world($world, $datdir, $batch_name, $CR, $reduction, $n_fluxons_wanted, $adapt, $force_make_world, $lim, $lim2, $configs, "initial");
 
     }
